Fixes (comments) fixed some models issues, so all basic endpoints work

+++ APIMMA/Dtos/PostDtos/PostDto.cs
@@ -11,7 +11,7 @@ namespace APIMMA.Dtos.PostDtos
 
         public string Type { get; set; } // MANUAL | TRAININGLOG | CHALLENGE
 
-        public BigInteger LikesCount { get; set; }
+        public long LikesCount { get; set; }
 
         public DateTime Created_at { get; set; }
         public UserSimplifiedDto User { get; set; }
